sdk/tests/conformance/extensions/oes-texture-half-float-null.html

<!--

/*
** Copyright (c) 2013 The Khronos Group Inc.
**
** Permission is hereby granted, free of charge, to any person obtaining a
** copy of this software and/or associated documentation files (the
** "Materials"), to deal in the Materials without restriction, including
** without limitation the rights to use, copy, modify, merge, publish,
** distribute, sublicense, and/or sell copies of the Materials, and to
** permit persons to whom the Materials are furnished to do so, subject to
** the following conditions:
**
** The above copyright notice and this permission notice shall be included
** in all copies or substantial portions of the Materials.
**
** THE MATERIALS ARE PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
** MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
** IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
** CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
** TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
** MATERIALS OR THE USE OR OTHER DEALINGS IN THE MATERIALS.
*/

-->

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>WebGL OES_texture_float Conformance Tests</title>
<link rel="stylesheet" href="../../resources/js-test-style.css"/>
<script src="../../resources/desktop-gl-constants.js" type="text/javascript"></script>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/webgl-test.js"></script>
<script src="../resources/webgl-test-utils.js"></script>
</head>
<body>
<div id="description"></div>
<canvas id="canvas" style="width: 50px; height: 50px;"> </canvas>
<div id="console"></div>
<script id="testFragmentShader" type="x-shader/x-fragment">
void main()
{
    gl_FragColor = vec4(0.0, 1.0, 0.0, 1.0);
}
</script>
<script>
"use strict"
description("This test verifies the functionality of OES_texture_half_float with null ArrayBufferView");

debug("");

var wtu = WebGLTestUtils;
var canvas = document.getElementById("canvas");
var gl = wtu.create3DContext(canvas);

if (!gl) {
    testFailed("WebGL context does not exists");
} else {
    testPassed("WebGL context exists");

    var shaders = [
        wtu.setupSimpleTextureVertexShader(gl),
        "testFragmentShader"
    ];
    var testProgram = wtu.setupProgram(gl,
                                       shaders,
                                       ['vPosition', 'texCoord0'],
                                       [0,1]);
    var quadParameters = wtu.setupUnitQuad(gl, 0, 1);

    // Verify that allocation of texture fails if extension is not enabled
    runTextureCreationTest(false);

    if(!gl.getExtension("OES_texture_half_float")) {
        testPassed("No OES_texture_half_float support. This is legal");
    } else {
        testPassed("Successfully enabled OES_texture_half_float extension");
        // Check if creation of texture succeed's with various formats
        runTextureCreationTest(testProgram, true, gl.RGBA);
        runTextureCreationTest(testProgram, true, gl.RGB);
        runTextureCreationTest(testProgram, true, gl.LUMINANCE);
        runTextureCreationTest(testProgram, true, gl.ALPHA);
        runTextureCreationTest(testProgram, true, gl.ALPHA);
        runTextureCreationTest(testProgram, true, gl.LUMINANCE_ALPHA);
        // Check if attaching texture as FBO target succeeds (Not mandatory)
        runRenderTargetTest(testProgram);
        // Check of getExtension() returns same object
        runUniqueObjectTest();
    }
}

function allocateTexture()
{
    var texture = gl.createTexture();
    gl.bindTexture(gl.TEXTURE_2D, texture);
    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST);
    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST);
    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
    glErrorShouldBe(gl, gl.NO_ERROR, "texture parameter setup should succeed");
    return texture;
}

function checkRenderingResults()
{
    wtu.checkCanvas(gl, [0, 255, 0, 255], "should be green");
}

function runTextureCreationTest(testProgram, extensionEnabled, format)
{
    debug("Testing texture creation with extension " + (extensionEnabled ? "enabled" : "disabled") +
          ". Expect " + (extensionEnabled ? "success" : "failure"));

    var texture = allocateTexture();
    var width = 2;
    var height = 2;
    gl.texImage2D(gl.TEXTURE_2D, 0, format, width, height, 0, format, gl.HALF_FLOAT_OES, null);
    if(!extensionEnabled) {
        glErrorShouldBe(gl, gl.INVALID_ENUM, "Half floating point texture must be diallowed if OES_texture_half_float isn't enabled");
        return;
    } else {
        glErrorShouldBe(gl, gl.NO_ERROR, "Half floating point texture allocation should succeed if OES_texture_half_float is enabled");
    }
    gl.uniform1i(gl.getUniformLocation(testProgram, "tex"), 0);
    wtu.clearAndDrawUnitQuad(gl);
    checkRenderingResults();
}

function runRenderTargetTest(testProgram)
{
    debug("");
    debug("Testing half floating point render target");

    var texture = allocateTexture();
    var width = 2;
    var height = 2;
    
    gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, width, height, 0, gl.RGBA, gl.HALF_FLOAT_OES, null);
    glErrorShouldBe(gl, gl.NO_ERROR, "Half floating point texture allocation should succeed if OES_texture_half_float is enabled");

    // Try to use this texture as render target
    var fbo = gl.createFramebuffer();
    gl.bindFramebuffer(gl.FRAMEBUFFER, fbo);
    gl.framebufferTexture2D(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.TEXTURE_2D, texture, 0);
    gl.bindTexture(gl.TEXTURE_2D, null);

    // It is legal for a WebGL implementation exposing the OES_texture_half_float extension to
    // support half floating point textures but not as attachments to framebuffer objects.
    if (gl.checkFramebufferStatus(gl.FRAMEBUFFER) != gl.FRAMEBUFFER_COMPLETE) {
        debug("Half floating point render targets not supported -- this is legal");
        return;
    }

    wtu.clearAndDrawUnitQuad(gl);
    glErrorShouldBe(gl, gl.NO_ERROR, "rendering to floating-point texture should succeed");

    // Now sample from the floating-point texture and verify we got the correct values.
    gl.bindFramebuffer(gl.FRAMEBUFFER, null);
    gl.bindTexture(gl.TEXTURE_2D, texture);
    gl.useProgram(testProgram);
    gl.uniform1i(gl.getUniformLocation(testProgram, "tex"), 0);
    wtu.clearAndDrawUnitQuad(gl);
    glErrorShouldBe(gl, gl.NO_ERROR, "rendering from half floating point texture should succeed");
    checkRenderingResults();
}

function runUniqueObjectTest()
{
    debug("Testing that getExtension() returns the same object each time");
    gl.getExtension("OES_texture_half_float").myProperty = 2;
    gc();
    shouldBe('gl.getExtension("OES_texture_half_float").myProperty', '2');
}

debug("");
var successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>

</body>
</html>
